Version 0.1 release
Major release which fully adds in the load and save system with some bug fixes and features in the backend. Licenses have more integration with the in game menus as well.

Version 0.1 release

Major release which fully adds in the load and save system with some bug fixes and features in the backend. Licenses have more integration with the in game menus as well.